requirements.txt
pyTelegramBotAPI
qrcode
Flask
reportlab
python-barcode
Werkzeug
gunicorn
